Refactor utils and add <Form>
