+++ packages/formik/src/index.tsx
@@ -7,3 +7,4 @@
 export * from './useFormik';
 export * from './useField';
 export * from './FormikContext';
+export * from './Form';
