--- packages/formik/src/utils.tsx
@@ -1,200 +1,0 @@
-/**
- * Copyright (c) Formik, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-import clone from 'lodash/clone';
-import toPath from 'lodash/toPath';
-import * as React from 'react';
-
-// Assertions
-
-/** @private is the value an empty array? */
-export const isEmptyArray = (value?: any) =>
-  Array.isArray(value) && value.length === 0;
-
-/** @private is the given object a Function? */
-export const isFunction = (obj: any): obj is Function =>
-  typeof obj === 'function';
-
-/** @private is the given object an Object? */
-export const isObject = (obj: any): obj is Object =>
-  obj !== null && typeof obj === 'object';
-
-/** @private is the given object an integer? */
-export const isInteger = (obj: any): boolean =>
-  String(Math.floor(Number(obj))) === obj;
-
-/** @private is the given object a string? */
-export const isString = (obj: any): obj is string =>
-  Object.prototype.toString.call(obj) === '[object String]';
-
-/** @private is the given object a NaN? */
-// eslint-disable-next-line no-self-compare
-export const isNaN = (obj: any): boolean => obj !== obj;
-
-/** @private Does a React component have exactly 0 children? */
-export const isEmptyChildren = (children: any): boolean =>
-  React.Children.count(children) === 0;
-
-/** @private is the given object/value a promise? */
-export const isPromise = (value: any): value is PromiseLike<any> =>
-  isObject(value) && isFunction(value.then);
-
-/** @private is the given object/value a type of synthetic event? */
-export const isInputEvent = (value: any): value is React.SyntheticEvent<any> =>
-  value && isObject(value) && isObject(value.target);
-
-/** @private Are we in RN? */
-export const isReactNative =
-  typeof window !== 'undefined' &&
-  window.navigator &&
-  window.navigator.product &&
-  window.navigator.product === 'ReactNative';
-
-/**
- * Same as document.activeElement but wraps in a try-catch block. In IE it is
- * not safe to call document.activeElement if there is nothing focused.
- *
- * The activeElement will be null only if the document or document body is not
- * yet defined.
- *
- * @param {?Document} doc Defaults to current document.
- * @return {Element | null}
- * @see https://github.com/facebook/fbjs/blob/master/packages/fbjs/src/core/dom/getActiveElement.js
- */
-export function getActiveElement(doc?: Document): Element | null {
-  doc = doc || (typeof document !== 'undefined' ? document : undefined);
-  if (typeof doc === 'undefined') {
-    return null;
-  }
-  try {
-    return doc.activeElement || doc.body;
-  } catch (e) {
-    return doc.body;
-  }
-}
-
-/**
- * Deeply get a value from an object via its path.
- */
-export function getIn(
-  obj: any,
-  key: string | string[],
-  def?: any,
-  p: number = 0
-) {
-  const path = toPath(key);
-  while (obj && p < path.length) {
-    obj = obj[path[p++]];
-  }
-  return obj === undefined ? def : obj;
-}
-
-/**
- * Deeply set a value from in object via it's path. If the value at `path`
- * has changed, return a shallow copy of obj with `value` set at `path`.
- * If `value` has not changed, return the original `obj`.
- *
- * Existing objects / arrays along `path` are also shallow copied. Sibling
- * objects along path retain the same internal js reference. Since new
- * objects / arrays are only created along `path`, we can test if anything
- * changed in a nested structure by comparing the object's reference in
- * the old and new object, similar to how russian doll cache invalidation
- * works.
- *
- * In earlier versions of this function, which used cloneDeep, there were
- * issues whereby settings a nested value would mutate the parent
- * instead of creating a new object. `clone` avoids that bug making a
- * shallow copy of the objects along the update path
- * so no object is mutated in place.
- *
- * Before changing this function, please read through the following
- * discussions.
- *
- * @see https://github.com/developit/linkstate
- * @see https://github.com/jaredpalmer/formik/pull/123
- */
-export function setIn(obj: any, path: string, value: any): any {
-  let res: any = clone(obj); // this keeps inheritance when obj is a class
-  let resVal: any = res;
-  let i = 0;
-  let pathArray = toPath(path);
-
-  for (; i < pathArray.length - 1; i++) {
-    const currentPath: string = pathArray[i];
-    let currentObj: any = getIn(obj, pathArray.slice(0, i + 1));
-
-    if (currentObj && (isObject(currentObj) || Array.isArray(currentObj))) {
-      resVal = resVal[currentPath] = clone(currentObj);
-    } else {
-      const nextPath: string = pathArray[i + 1];
-      resVal = resVal[currentPath] =
-        isInteger(nextPath) && Number(nextPath) >= 0 ? [] : {};
-    }
-  }
-
-  // Return original object if new value is the same as current
-  if ((i === 0 ? obj : resVal)[pathArray[i]] === value) {
-    return obj;
-  }
-
-  if (value === undefined) {
-    delete resVal[pathArray[i]];
-  } else {
-    resVal[pathArray[i]] = value;
-  }
-
-  // If the path array has a single element, the loop did not run.
-  // Deleting on `resVal` had no effect in this scenario, so we delete on the result instead.
-  if (i === 0 && value === undefined) {
-    delete res[pathArray[i]];
-  }
-
-  return res;
-}
-
-// React currently throws a warning when using useLayoutEffect on the server.
-// To get around it, we can conditionally useEffect on the server (no-op) and
-// useLayoutEffect in the browser.
-// @see https://gist.github.com/gaearon/e7d97cdf38a2907924ea12e4ebdf3c85
-const useIsomorphicLayoutEffect =
-  typeof window !== 'undefined' &&
-  typeof window.document !== 'undefined' &&
-  typeof window.document.createElement !== 'undefined'
-    ? React.useLayoutEffect
-    : React.useEffect;
-
-export function useEventCallback<T extends (...args: any[]) => any>(fn: T): T {
-  const ref: any = React.useRef(fn);
-
-  // we copy a ref to the callback scoped to the current state/props on each render
-  useIsomorphicLayoutEffect(() => {
-    ref.current = fn;
-  });
-
-  return React.useCallback(
-    (...args: any[]) => ref.current.apply(void 0, args),
-    []
-  ) as T;
-}
-
-export function useStateAndRef<T>(
-  initialValue: T
-): [T, React.Dispatch<React.SetStateAction<T>>, React.MutableRefObject<T>] {
-  const ref = React.useRef<T>(initialValue);
-  const [state, setState] = React.useState<T>(initialValue);
-  const update = useEventCallback(v => {
-    ref.current = v;
-    setState(v);
-  });
-  return [state, update, ref];
-}
-
-const noopReducer = (s: number) => s + 1;
-export function useForceRender() {
-  const [, forceRender] = React.useReducer(noopReducer, 0);
-  return forceRender;
-}
