+++ packages/formik/src/utils/getIn.ts
@@ -1,0 +1,22 @@
+/**
+ * Copyright (c) Formik, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+import toPath from 'lodash/toPath';
+/**
+ * Deeply get a value from an object via its path.
+ */
+export function getIn(
+  obj: any,
+  key: string | string[],
+  def?: any,
+  p: number = 0
+) {
+  const path = toPath(key);
+  while (obj && p < path.length) {
+    obj = obj[path[p++]];
+  }
+  return obj === undefined ? def : obj;
+}
